+++ src/ts/Ads/Utilities/ProgrammaticTrackingService.ts
@@ -2,6 +2,7 @@ import { Platform } from 'Core/Constants/Platform';
 import { INativeResponse, RequestManager } from 'Core/Managers/RequestManager';
 import { ClientInfo } from 'Core/Models/ClientInfo';
 import { DeviceInfo } from 'Core/Models/DeviceInfo';
+import { CustomFeatures } from './CustomFeatures';
 
 export enum ProgrammaticTrackingError {
     TooLargeFile = 'too_large_file', // a file 20mb and over are considered too large
@@ -41,14 +42,17 @@ export enum VastMetric {
 
 export enum MiscellaneousMetric {
     CampaignNotFound = 'campaign_not_found',
-    ConsentParagraphLinkClicked = 'consent_paragraph_link_clicked'
+    ConsentParagraphLinkClicked = 'consent_paragraph_link_clicked',
+    CampaignAttemptedToShowInBackground = 'ad_attempted_show_background'
 }
 
 export enum LoadMetric {
     LoadEnabledAuctionRequest = 'load_enabled_auction_request',
     LoadEnabledFill = 'load_enabled_fill',
     LoadEnabledNoFill = 'load_enabled_no_fill',
-    LoadEnabledShow = 'load_enabled_show'
+    LoadEnabledShow = 'load_enabled_show',
+    LoadEnabledInitializationSuccess = 'load_enabled_initialization_success',
+    LoadEnabledInitializationFailure = 'load_enabled_initialization_failure'
 }
 
 type ProgrammaticTrackingMetric = AdmobMetric | BannerMetric | ChinaMetric | VastMetric | MiscellaneousMetric | LoadMetric;
@@ -116,6 +120,12 @@ export class ProgrammaticTrackingService {
     }
 
     public reportMetric(event: ProgrammaticTrackingMetric): Promise<INativeResponse> {
+
+        const isLoadMetric = event in LoadMetric;
+        const isZyngaSolitareUsingLoad = (CustomFeatures.isTrackedGameUsingLoadApi(this._clientInfo.getGameId()));
+        if (isLoadMetric && !(isZyngaSolitareUsingLoad)) {
+            return Promise.resolve(<INativeResponse>{});
+        }
         const metricData: IProgrammaticTrackingData = {
             metrics: [
                 {